+++ libqtile/backend/wayland/__init__.py
@@ -17,11 +17,3 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
-try:
-    # Shorten the import for this because it will be used in configs
-    from libqtile.backend.wayland.inputs import InputConfig  # noqa: F401
-except ModuleNotFoundError:
-    print("InputConfig couldn't be imported from libqtile.backend.wayland")
-    print("If this happened during setup.py installation, ignore this message.")
-    print("Otherwise, make sure to run ./libqtile/backend/wayland/cffi/build.py")
